--- packages/frontend/src/components/RegisterForms/VerifyEmailForm.tsx
@@ -1,128 +1,0 @@
-import {
-    Anchor,
-    Button,
-    PinInput,
-    Stack,
-    Text,
-    Title,
-    UnstyledButton,
-} from '@mantine/core';
-import { isNotEmpty, useForm } from '@mantine/form';
-import React, { FC, useEffect, useState } from 'react';
-import Countdown, { zeroPad } from 'react-countdown';
-import {
-    useEmailStatus,
-    useOneTimePassword,
-    useVerifyEmail,
-} from '../../hooks/useEmailVerification';
-import { useApp } from '../../providers/AppProvider';
-import LoadingState from '../common/LoadingState';
-
-const VerifyEmailForm: FC<{ isLoading?: boolean }> = ({ isLoading }) => {
-    const { health, user } = useApp();
-    const { mutate: verifyCode, isLoading: verificationLoading } =
-        useVerifyEmail();
-    const { data, isLoading: statusLoading } = useEmailStatus();
-    const { mutate: sendVerificationEmail, isLoading: emailLoading } =
-        useOneTimePassword();
-    const form = useForm<{ code: number }>({
-        initialValues: {
-            code: 0,
-        },
-        validate: {
-            code: isNotEmpty('This field is required.'),
-        },
-    });
-    const errorMessage = form.errors.code;
-    const expirationTime = data?.otp?.expiresAt || new Date();
-    const loadingState =
-        statusLoading || emailLoading || health.isLoading || isLoading;
-
-    useEffect(() => {
-        if (data?.otp && data?.otp.numberOfAttempts > 0) {
-            const remainingAttempts = 5 - data.otp.numberOfAttempts;
-            const message = data.otp.isExpired
-                ? 'Your one-time password expired. Please resend a verification email.'
-                : data.otp.numberOfAttempts < 5
-                ? `The code doesn't match the one we sent you. You have ${remainingAttempts} attempt${
-                      remainingAttempts > 1 ? 's' : ''
-                  } left.`
-                : "Hmm that code doesn't match the one we sent you. You've already had 5 attempts, please resend a verification email and try again.";
-            form.setFieldError('code', message);
-        } else {
-            form.clearFieldError('code');
-        }
-    }, [data, form.setFieldError, form.clearFieldError]);
-
-    if (loadingState) {
-        return <LoadingState title="" />;
-    }
-
-    return (
-        // FIXME: use Mantine sizes for width
-        <Stack spacing="md" justify="center" align="center" w={300} mx="auto">
-            <Title order={3}>Check your inbox!</Title>
-            <Text color="gray.6" ta="center">
-                Verify your email address by entering the code we've just sent
-                to <b>{user?.data?.email || 'your email'}</b>
-            </Text>
-            <form
-                name="verifyEmail"
-                onSubmit={form.onSubmit((values: { code: number }) =>
-                    verifyCode(values.code.toString()),
-                )}
-            >
-                <Stack spacing="xs" justify="center" align="center" mt="md">
-                    <PinInput
-                        aria-label="One-time password"
-                        name="code"
-                        length={6}
-                        oneTimeCode
-                        disabled={data?.otp?.isMaxAttempts}
-                        {...form.getInputProps('code')}
-                    />
-                    <Text ta="center" color="red.7">
-                        {errorMessage?.toString()}
-                    </Text>
-                </Stack>
-                <Countdown
-                    key={expirationTime?.toString()}
-                    date={expirationTime}
-                    renderer={({ minutes, seconds, completed }) => {
-                        if (completed || data?.otp?.isMaxAttempts) {
-                            return <></>;
-                        }
-                        return (
-                            // FIXME: use Mantine sizes for width
-                            <Stack spacing="xs" mt="md" w={290} align="center">
-                                <Button
-                                    loading={verificationLoading}
-                                    type="submit"
-                                    w={230}
-                                >
-                                    Submit
-                                </Button>
-                                <Text color="gray.6" ta="center">
-                                    Your one-time password expires in{' '}
-                                    <b>
-                                        {zeroPad(minutes)}:{zeroPad(seconds)}
-                                    </b>
-                                </Text>
-                            </Stack>
-                        );
-                    }}
-                />
-            </form>
-            <UnstyledButton
-                onClick={() => {
-                    form.reset();
-                    sendVerificationEmail();
-                }}
-            >
-                <Anchor size="sm">Resend email</Anchor>
-            </UnstyledButton>
-        </Stack>
-    );
-};
-
-export default VerifyEmailForm;
