+++ packages/frontend/src/pages/Register.tsx
@@ -10,8 +10,8 @@ import {
     OneLoginLoginButton,
 } from '../components/common/GoogleLoginButton';
 import Page from '../components/common/Page/Page';
+import CreateUserForm from '../components/CreateUserForm';
 import PageSpinner from '../components/PageSpinner';
-import CreateUserForm from '../components/RegisterForms/CreateUserForm';
 import useToaster from '../hooks/toaster/useToaster';
 import { useApp } from '../providers/AppProvider';
 import { useTracking } from '../providers/TrackingProvider';
