+++ packages/frontend/src/pages/VerifyEmail.tsx
@@ -15,9 +15,9 @@ import { Helmet } from 'react-helmet';
 import { useHistory } from 'react-router-dom';
 import { useIntercom } from 'react-use-intercom';
 import Page from '../components/common/Page/Page';
+import VerifyEmailForm from '../components/CreateUserForm/VerifyEmailForm';
 import PageSpinner from '../components/PageSpinner';
 import { StyledSuccessIcon } from '../components/ProjectConnection/ProjectConnectFlow/ProjectConnectFlow.styles';
-import VerifyEmailForm from '../components/RegisterForms/VerifyEmailForm';
 import { useEmailStatus } from '../hooks/useEmailVerification';
 import { useApp } from '../providers/AppProvider';
 import LightdashLogo from '../svgs/lightdash-black.svg';
